api/serializers.py
from django.contrib.auth.models import User, Group
from rest_framework import serializers

from helpline.models import HelplineUser, Case

class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ('url', 'username', 'email', 'groups')

class GroupSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Group
        fields = ('url', 'name')

class HelplineUserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = HelplineUser
        fields = ('url', 'user', 'chlauth', 'chlexten', 'chlcase', 'chlstatus')

class HelplineCaseSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Case
        fields = ('url', 'chlkey', 'chltime', 'chlpriority', 'chldata', 'chlpopup')
