requirements.txt
beautifulsoup4==4.6.0
click==6.7
croniter==0.3.20
Django==1.10.1
django-admin-tools==0.8.0
django-admin-tools-stats==0.7.1
django-appconf==1.0.2
django-avatar==3.1.0
django-bootstrap-form==3.3
django-bower==5.2.0
django-cache-utils==0.7.2
django-cors-headers==2.2.0
django-countries==4.6.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.4
git+git://github.com/arthanson/django-genericadmin.git@master#egg=genericadmin
django-helpdesk==0.2.7
-e git+https://github.com/Kirembu/django-helpline-faq.git@master#egg=faq
django-jsonfield==1.0.1
django-jsonview==1.0.0
django-mailer==1.2.2
django-markdown-deux==1.0.5
django-model-utils==3.1.1
-e git+https://github.com/django-notifications/django-notifications.git@2d22f44000b9c7b45fcc0110bb959121cfc9522a#egg=django_notifications_hq
django-nvd3==0.9.7
django-qsstats-magic==1.0.0
django-report-tools==0.2.2
django-selectable==0.9.0
-e git+https://github.com/kirembu/django-sendsms@e69ace26750bb8cc6dcfdc49412d6d6c2ea26733#egg=django_sendsms
django-tables2==1.21.1
django-tables2-reports==0.0.10
django.js==0.8.1
djangorestframework==3.7.7
docutils==0.14
email-reply-parser==0.5.9
et-xmlfile==1.0.1
feedparser==5.2.1
gunicorn==19.7.1
iptools==0.6.1
itsdangerous==0.24
jdcal==1.3
Jinja2==2.10
jsonfield==2.0.2
lockfile==0.12.2
lxml==4.1.1
markdown2==2.3.5
MarkupSafe==0.23
MySQL-python==1.2.5
mysqlclient==1.3.9
nameparser==0.5.6
odfpy==1.3.6
openpyxl==2.5.1
path.py==8.2.1
pbr==1.8.1
Pillow==3.3.0
psycopg2-binary==2.7.5
py-Asterisk==0.5.16
pycall==2.2
python-dateutil==2.5.3
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.10.0
rq==0.10.0
rq-scheduler==0.6.1
simplejson==3.13.2
six==1.10.0
sqlparse==0.1.19
stevedore==1.9.0
tablib==0.12.1
unicodecsv==0.14.1
Unidecode==0.4.19
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
xlrd==1.1.0
xlwt==1.1.2
